flask/requirements.txt
autopep8==1.4.4
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
fake-useragent==0.1.11
feedparser==5.2.1
Flask==1.1.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
numpy==1.17.2
pandas==0.25.1
pycodestyle==2.5.0
PyMySQL==0.9.3
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.2
requests==2.22.0
six==1.12.0
soupsieve==1.9.4
SQLAlchemy==1.3.8
urllib3==1.25.6
Werkzeug==0.15.6
WTForms==2.2.1
uWSGI==2.0.18